Split GUI helper tests by function
